+++ fehu/lib/vector_env.ml
@@ -192,4 +192,8 @@ let step vector_env actions =
   let infos = Array.map (fun (_, _, _, _, info) -> info) results in
   { observations; rewards; terminations; truncations; infos }
 
+let render vector_env = Array.map Env.render vector_env.envs
+
+let envs vector_env = vector_env.envs
+
 let close vector_env = Array.iter Env.close vector_env.envs